feat: Refactor booking page and add login page structure
