+++ terrainapp/webapp/src/main.jsx
@@ -1,10 +1,14 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import App from './App.jsx'
-import './index.css'
+// src/main.jsx
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
+import BookingPage from './BookingPage.jsx'; // Import the BookingPage component
+import Login from './Login.jsx'; // Import the Login component
+import './index.css';
 
+
+// Change BookingPage to login to view the login page and vice versa 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <App />
+    <Login />
   </StrictMode>,
-)
+);
